How_r_u_AI/requirements.txt
pip==22.0.2
setuptools==59.6.0
Flask>=2.3.2
torch>=2.0.1
torchvision>=0.15.2
ultralytics>=8.0.0
opencv-python
Pillow
